webpki
Module alg_id

Encodings of the PKIX AlgorithmIdentifier type.

This module contains a set of common values, and exists to keep the names of these separate from the actual algorithm implementations.

Constants

ECDSA_P256

AlgorithmIdentifier for id-ecPublicKey with named curve secp256r1.

ECDSA_P384

AlgorithmIdentifier for id-ecPublicKey with named curve secp384r1.

ECDSA_P521

AlgorithmIdentifier for id-ecPublicKey with named curve secp521r1.

ECDSA_SHA256

AlgorithmIdentifier for ecdsa-with-SHA256.

ECDSA_SHA384

AlgorithmIdentifier for ecdsa-with-SHA384.

ECDSA_SHA512

AlgorithmIdentifier for ecdsa-with-SHA512.

ED25519

AlgorithmIdentifier for ED25519.

RSA_ENCRYPTION

AlgorithmIdentifier for rsaEncryption.

RSA_PKCS1_SHA256

AlgorithmIdentifier for sha256WithRSAEncryption.

RSA_PKCS1_SHA384

AlgorithmIdentifier for sha384WithRSAEncryption.

RSA_PKCS1_SHA512

AlgorithmIdentifier for sha512WithRSAEncryption.

RSA_PSS_SHA256

AlgorithmIdentifier for rsassaPss with:

RSA_PSS_SHA384

AlgorithmIdentifier for rsassaPss with:

RSA_PSS_SHA512

AlgorithmIdentifier for rsassaPss with: