webpki

Module crl

Re-exports

pub use types::BorrowedCertRevocationList;

pub use types::BorrowedRevokedCert;

pub use types::CertRevocationList;

pub use types::RevocationReason;

pub use types::OwnedCertRevocationList;

pub use types::OwnedRevokedCert;

Modules

types 🔒

Structs

CertNotRevoked 🔒

CrlsRequired

An opaque error indicating the caller must provide at least one CRL when building a RevocationOptions instance.

RevocationOptions

Describes how revocation checking is performed, if at all. Can be constructed with a RevocationOptionsBuilder instance.

RevocationOptionsBuilder

Builds a RevocationOptions instance to control how revocation checking is performed.

Enums

ExpirationPolicy

Describes how to handle the nextUpdate field of the CRL (i.e. expiration).

KeyUsageMode 🔒

RevocationCheckDepth

Describes how much of a certificate chain is checked for revocation status.

UnknownStatusPolicy

Describes how to handle the case where a certificate’s revocation status is unknown.

Functions

crl_signature_err 🔒