Enum webpki::CertRevocationList

pub enum CertRevocationList<'a> {
    Owned(OwnedCertRevocationList),
    Borrowed(BorrowedCertRevocationList<'a>),
}

A RFC 52801 profile Certificate Revocation List (CRL).

May be either an owned, or a borrowed representation.

https://www.rfc-editor.org/rfc/rfc5280#section-5 ↩

Variants

Owned(OwnedCertRevocationList)

An owned representation of a CRL.

Borrowed(BorrowedCertRevocationList<'a>)

A borrowed representation of a CRL.

Implementations

impl<'a> CertRevocationList<'a>

pub fn issuer(&self) -> &[u8]

Return the DER encoded issuer of the CRL.

pub fn issuing_distribution_point(&self) -> Option<&[u8]>

Return the DER encoded issuing distribution point of the CRL, if any.

pub fn find_serial( &self, serial: &[u8], ) -> Result<Option<BorrowedRevokedCert<'_>>, Error>

Try to find a revoked certificate in the CRL by DER encoded serial number. This may yield an error if the CRL has malformed revoked certificates.