webpki

Module signed_data

Structs

OwnedSignedData 🔒

X.509 certificates and related items that are signed are almost always encoded in the format “tbs||signatureAlgorithm||signature”. This structure captures this pattern as an owned data type.

SignedData 🔒

X.509 certificates and related items that are signed are almost always encoded in the format “tbs||signatureAlgorithm||signature”. This structure captures this pattern.

SubjectPublicKeyInfo 🔒

Functions

verify_signature 🔒

verify_signed_data 🔒

Verify signed_data using the public key in the DER-encoded SubjectPublicKeyInfo spki using one of the algorithms in supported_algorithms.