Struct webpki::DNSName

pub struct DNSName(_);

A DNS Name suitable for use in the TLS Server Name Indication (SNI) extension and/or for use as the reference hostname for which to verify a certificate.

A DNSName is guaranteed to be syntactically valid. The validity rules are specified in RFC 5280 Section 7.2, except that underscores are also allowed.

DNSName stores a copy of the input it was constructed from in a String and so it is only available when the std default feature is enabled.

Eq, PartialEq, etc. are not implemented because name comparison frequently should be done case-insensitively and/or with other caveats that depend on the specific circumstances in which the comparison is done.

Methods

impl DNSName

fn as_ref(&self) -> DNSNameRef

Returns a DNSNameRef that refers to this DNSName.