Struct webpki::DNSNameRef

pub struct DNSNameRef<'a>(_);

A reference to a DNS Name suitable for use in the TLS Server Name Indication (SNI) extension and/or for use as the reference hostname for which to verify a certificate.

A DNSNameRef is guaranteed to be syntactically valid. The validity rules are specified in RFC 5280 Section 7.2, except that underscores are also allowed.

Eq, PartialEq, etc. are not implemented because name comparison frequently should be done case-insensitively and/or with other caveats that depend on the specific circumstances in which the comparison is done.

Methods

impl<'a> DNSNameRef<'a>

fn try_from_ascii(dns_name: Input<'a>) -> Result<Self, ()>

Constructs a DNSNameRef from the given input if the input is a syntactically-valid DNS name.

fn try_from_ascii_str(dns_name: &str) -> Result<DNSNameRef, ()>

Constructs a DNSNameRef from the given input if the input is a syntactically-valid DNS name.