webpki

Trait ExtendedKeyUsageValidator

pub trait ExtendedKeyUsageValidator {
    // Required method
    fn validate(&self, iter: KeyPurposeIdIter<'_, '_>) -> Result<(), Error>;
}

A trait for validating the Extended Key Usage (EKU) extensions of a certificate.

Required Methods

fn validate(&self, iter: KeyPurposeIdIter<'_, '_>) -> Result<(), Error>

Validate the EKU values in a certificate.

iter yields the EKU OIDs in the certificate, or an error if the EKU extension is malformed. validate() should yield Ok(()) if the EKU values match the required policy, or an Error if they do not. Ideally the Error should be Error::RequiredEkuNotFoundContext if the policy is not met.