Function webpki::trust_anchor_util::generate_code_for_trust_anchors

pub fn generate_code_for_trust_anchors(
    name: &str, 
    trust_anchors: &[TrustAnchor<'_>]
) -> String

Generates code for hard-coding the given trust anchors into a program. This is designed to be used in a build script. name is the name of the public static variable that will contain the TrustAnchor array.