Type Alias webpki::DNSNameRef

pub type DNSNameRef<'a> = DnsNameRef<'a>;

👎Deprecated: use DnsNameRef

Aliased Type

struct DNSNameRef<'a>(/* private fields */);

Implementations

impl<'a> DnsNameRef<'a>

pub fn try_from_ascii(dns_name: &'a [u8]) -> Result<Self, InvalidDnsNameError>

Constructs a DnsNameRef from the given input if the input is a syntactically-valid DNS name.

pub fn try_from_ascii_str( dns_name: &'a str ) -> Result<Self, InvalidDnsNameError>

Constructs a DnsNameRef from the given input if the input is a syntactically-valid DNS name.

pub fn to_owned(&self) -> DnsName

Constructs a DnsName from this DnsNameRef

Requires the alloc feature.