webpki::x509

Enum DistributionPointName

pub(crate) enum DistributionPointName<'a> {
    NameRelativeToCrlIssuer,
    FullName(DerIterator<'a, GeneralName<'a>>),
}

A certificate revocation list (CRL) distribution point name, describing a source of CRL information for a given certificate as described in RFC 5280 section 4.2.3.131.

https://datatracker.ietf.org/doc/html/rfc5280#section-4.2.1.13 ↩

Variants

NameRelativeToCrlIssuer

The distribution point name is a relative distinguished name, relative to the CRL issuer.

FullName(DerIterator<'a, GeneralName<'a>>)

The distribution point name is a sequence of GeneralName items.