Crate webpki

webpki: Web PKI X.509 Certificate Validation.

See EndEntityCert’s documentation for a description of the certificate processing steps necessary for a TLS connection.

Features

Feature	Description

alloc	Enable features that require use of the heap. Currently all RSA signature algorithms require this feature.

std	Enable features that require libstd. Implies alloc.

Structs

DnsNameRef

A reference to a DNS Name suitable for use in the TLS Server Name Indication (SNI) extension and/or for use as the reference hostname for which to verify a certificate.

EndEntityCert

An end-entity certificate.

InvalidDnsNameError

An error indicating that a DnsNameRef could not built because the input is not a syntactically-valid DNS Name.

SignatureAlgorithm

A signature algorithm.

Time

The time type.

TlsClientTrustAnchors

Trust anchors which may be used for authenticating clients.

TlsServerTrustAnchors

Trust anchors which may be used for authenticating servers.

TrustAnchor

A trust anchor (a.k.a. root CA).

Enums

Error

An error that occurs during certificate validation or name validation.

Statics

ECDSA_P256_SHA256

ECDSA signatures using the P-256 curve and SHA-256.

ECDSA_P256_SHA384

ECDSA signatures using the P-256 curve and SHA-384. Deprecated.

ECDSA_P384_SHA256

ECDSA signatures using the P-384 curve and SHA-256. Deprecated.

ECDSA_P384_SHA384

ECDSA signatures using the P-384 curve and SHA-384.

ED25519

ED25519 signatures according to RFC 8410

Type Aliases

DNSNameRefDeprecated

TLSClientTrustAnchorsDeprecated

TLSServerTrustAnchorsDeprecated