Crate webpki

webpki: Web PKI X.509 Certificate Validation.

See EndEntityCert’s documentation for a description of the certificate processing steps necessary for a TLS connection.

Modules

trust_anchor_util

Utilities for efficiently embedding trust anchors in programs.

Structs

DNSName

A DNS Name suitable for use in the TLS Server Name Indication (SNI) extension and/or for use as the reference hostname for which to verify a certificate.

DNSNameRef

A reference to a DNS Name suitable for use in the TLS Server Name Indication (SNI) extension and/or for use as the reference hostname for which to verify a certificate.

EndEntityCert

An end-entity certificate.

InvalidDNSNameError

An error indicating that a DNSNameRef could not built because the input is not a syntactically-valid DNS Name.

SignatureAlgorithm

A signature algorithm.

TLSClientTrustAnchors

Trust anchors which may be used for authenticating clients.

TLSServerTrustAnchors

Trust anchors which may be used for authenticating servers.

Time

The time type.

TrustAnchor

A trust anchor (a.k.a. root CA).

Enums

Error

An error that occurs during certificate validation or name validation.

Statics

ECDSA_P256_SHA256

ECDSA signatures using the P-256 curve and SHA-256.

ECDSA_P256_SHA384

ECDSA signatures using the P-256 curve and SHA-384. Deprecated.

ECDSA_P384_SHA256

ECDSA signatures using the P-384 curve and SHA-256. Deprecated.

ECDSA_P384_SHA384

ECDSA signatures using the P-384 curve and SHA-384.

ED25519

ED25519 signatures according to RFC 8410

RSA_PKCS1_2048_8192_SHA256

RSA PKCS#1 1.5 signatures using SHA-256 for keys of 2048-8192 bits.

RSA_PKCS1_2048_8192_SHA384

RSA PKCS#1 1.5 signatures using SHA-384 for keys of 2048-8192 bits.

RSA_PKCS1_2048_8192_SHA512

RSA PKCS#1 1.5 signatures using SHA-512 for keys of 2048-8192 bits.

RSA_PKCS1_3072_8192_SHA384

RSA PKCS#1 1.5 signatures using SHA-384 for keys of 3072-8192 bits.

RSA_PSS_2048_8192_SHA256_LEGACY_KEY

RSA PSS signatures using SHA-256 for keys of 2048-8192 bits and of type rsaEncryption; see RFC 4055 Section 1.2.

RSA_PSS_2048_8192_SHA384_LEGACY_KEY

RSA PSS signatures using SHA-384 for keys of 2048-8192 bits and of type rsaEncryption; see RFC 4055 Section 1.2.

RSA_PSS_2048_8192_SHA512_LEGACY_KEY

RSA PSS signatures using SHA-512 for keys of 2048-8192 bits and of type rsaEncryption; see RFC 4055 Section 1.2.