Crate webpki

webpki: Web PKI X.509 Certificate Validation.

git clone https://github.com/briansmith/webpki

See EndEntityCert's documentation for a description of the certificate processing steps necessary for a TLS connection.

Modules

trust_anchor_util
Utilities for efficiently embedding trust anchors in programs.

Structs

EndEntityCert
An end-entity certificate.

SignatureAlgorithm
A signature algorithm.

TLSServerTrustAnchors
Trust anchors which may be used for authenticating servers.

Time
The time type.

TrustAnchor
A trust anchor (a.k.a. root CA).

Enums

Error
An error that occurs during certificate validation or name validation.

Statics

ECDSA_P256_SHA256
ECDSA signatures using the P-256 curve and SHA-256.

ECDSA_P256_SHA384
ECDSA signatures using the P-256 curve and SHA-384. Deprecated.

ECDSA_P384_SHA256
ECDSA signatures using the P-384 curve and SHA-256. Deprecated.

ECDSA_P384_SHA384
ECDSA signatures using the P-384 curve and SHA-384.

RSA_PKCS1_2048_8192_SHA1
RSA PKCS#1 1.5 signatures using SHA-1 for keys of 2048-8192 bits. Deprecated.

RSA_PKCS1_2048_8192_SHA256
RSA PKCS#1 1.5 signatures using SHA-256 for keys of 2048-8192 bits.

RSA_PKCS1_2048_8192_SHA384
RSA PKCS#1 1.5 signatures using SHA-384 for keys of 2048-8192 bits.

RSA_PKCS1_2048_8192_SHA512
RSA PKCS#1 1.5 signatures using SHA-512 for keys of 2048-8192 bits.

RSA_PKCS1_3072_8192_SHA384
RSA PKCS#1 1.5 signatures using SHA-384 for keys of 3072-8192 bits.

RSA_PSS_2048_8192_SHA256_LEGACY_KEY
RSA PSS signatures using SHA-256 for keys of 2048-8192 bits and of type rsaEncryption; see https://tools.ietf.org/html/rfc4055#section-1.2

RSA_PSS_2048_8192_SHA384_LEGACY_KEY
RSA PSS signatures using SHA-384 for keys of 2048-8192 bits and of type rsaEncryption; see https://tools.ietf.org/html/rfc4055#section-1.2

RSA_PSS_2048_8192_SHA512_LEGACY_KEY
RSA PSS signatures using SHA-512 for keys of 2048-8192 bits and of type rsaEncryption; see https://tools.ietf.org/html/rfc4055#section-1.2