webpki::signed_data

Function verify_signed_data

pub(crate) fn verify_signed_data(
    supported_algorithms: &[&dyn SignatureVerificationAlgorithm],
    spki_value: Input<'_>,
    signed_data: &SignedData<'_>,
    budget: &mut Budget,
) -> Result<(), Error>

Verify signed_data using the public key in the DER-encoded SubjectPublicKeyInfo spki using one of the algorithms in supported_algorithms.

The algorithm is chosen based on the algorithm information encoded in the algorithm identifiers in public_key and signed_data.algorithm. The ordering of the algorithms in supported_algorithms does not really matter, but generally more common algorithms should go first, as it is scanned linearly for matches.