webpki

Struct Cert

pub struct Cert<'a> { /* private fields */ }

A parsed X509 certificate.

Implementations

impl<'a> Cert<'a>

pub fn valid_dns_names(&self) -> impl Iterator<Item = &str>

Returns a list of valid DNS names provided in the subject alternative names extension

This function must not be used to implement custom DNS name verification. Checking that a certificate is valid for a given subject name should always be done with EndEntityCert::verify_is_valid_for_subject_name.

pub fn serial(&self) -> &[u8]

Raw DER encoded certificate serial number.

pub fn issuer(&self) -> &[u8]

Raw DER encoded certificate issuer.

pub fn subject(&self) -> &[u8]

Raw DER encoded certificate subject.

pub fn subject_public_key_info(&self) -> SubjectPublicKeyInfoDer<'static>

Get the RFC 5280-compliant SubjectPublicKeyInfoDer (SPKI) of this Cert.

pub fn der(&self) -> CertificateDer<'a>

Raw DER encoded representation of the certificate.