Module webpki::subject_name::dns_name

Structs

DnsName

A DNS Name suitable for use in the TLS Server Name Indication (SNI) extension and/or for use as the reference hostname for which to verify a certificate.

DnsNameRef

A reference to a DNS Name suitable for use in the TLS Server Name Indication (SNI) extension and/or for use as the reference hostname for which to verify a certificate.

InvalidDnsNameError

An error indicating that a DnsNameRef could not built because the input is not a syntactically-valid DNS Name.

WildcardDnsNameRef

A reference to a DNS Name presented by a server that may include a wildcard.

Enums

AllowWildcards 🔒

GeneralDnsNameRef

A DNS name that may be either a DNS name identifier presented by a server (which may include wildcards), or a DNS name identifier referenced by a client for matching purposes (wildcards not permitted).

IdRole 🔒

Functions

ascii_lower 🔒

is_valid_dns_id 🔒

presented_id_matches_constraint 🔒

presented_id_matches_reference_id 🔒

presented_id_matches_reference_id_internal 🔒